Added sample

+++ plugin/src/main/groovy/com/github/konifar/gradle/remover/remover/filetype/FileRemover.groovy
@@ -33,6 +33,11 @@ class FileRemover extends AbstractRemover {
         boolean isMatched = checkTargetTextMatches(extractFileName(file))
 
         if (!isMatched) {
+            if (isMatchedExcludeNames(file.path)) {
+                ColoredLogger.logYellow "[${fileType}]   Ignore to remove ${file.name}"
+                return false
+            }
+
             ColoredLogger.logGreen("[${fileType}]   Remove ${file.name}")
             if (!dryRun) {
                 file.delete()
